// src/HomePage/components/types.ts
export interface Tourz {
    icon: string;
    title: string;
    description?: string;
    nameClass?: string
}
export interface Review {
    icon: string;
    name: string;
    description: string;
    typeService: string
}
export interface Destination {
    image: string;
    name: string;
    slug: string;
    tours_count: string;
}
export interface Testimonials {
    name: string;
    position: string;
    avatar: string;
    title: string;
    message: string;
}